+++ middlewares/validators/firebase.js
@@ -1,5 +1,5 @@
 const Joi = require('joi');
-const { validError } = require('../errors/customError');
+const { validError } = require('../../utils/errors/customError');
 
 const resetPassword = async (req, res, next) => {
   const paramSchema = Joi.object({
